Use Module#include instead of alias_method (#3)
Since the time this gem was initially released, I came to realize that using `include` is a more friendly way to extend base functionality. Additionally it simplifies debugging and doesn't pollute the method list of the target.

Futhermore, Rails' `ClassMethods` paradigm didn't stick, so I'm switching to `Module#include self` to add "class methods" to the `LiterateMaruku` module.